### 2019-12-13 #1947 @ryEllison
I’d love to review this article.

Ryan

Sent from my iPhone

> On Dec 13, 2019, at 3:15 PM, Yuan Tang <notifications@github.com> wrote:
>
> ﻿
> 👋 @axel-loewe, @nan2stat, @dirmeier, @faroit, @JesseLivezey, @ryEllison, @b-freitas, @iskandr, @osolari, @dvanic, @innat, @aozorahime, @snikumbh, @gavinmdouglas Hi! If you'd like to review for this submission, please let us know here!
>
