### 2019-12-13 #1957 @whedon
[REVIEW]: scikit-hubness: Hubness Reduction and Approximate Neighbor Search
**Submitting author:** @VarIr (<a href="http://orcid.org/0000-0003-2216-4295">Roman Feldbauer</a>)
**Repository:** <a href="https://github.com/VarIr/scikit-hubness/" target ="_blank">https://github.com/VarIr/scikit-hubness/</a>
**Version:** v0.21.2
