### 2020-01-08 #1957 @arfon
> @terrytangyuan The missing DOI is for an article from arXiv, which does not provide DOIs, afaik. What should I do about this?

@VarIr - perhaps you can update your reference to what looks to be the (journal) published version of the same article. This way, the authors will receive citation credit from you. BTW, this comment assumes that the arXiv version is the same paper as https://doi.org/10.1109/tpami.2018.2889473 which during my very brief inspection appears to be the case.
