### 2020-01-08 #1957 @VarIr
Thank you for the hint! Indeed, I missed that it was eventually published in a journal.
