### 2020-01-09 #1957 @ryEllison
I'm satisfied that the requirements for publication have been met, and as such, I'm happy to see this accepted and published as submitted.

The documentation is extensive and functionality is exemplified both in text and through example usages. In an odd confluence of events, my lab has recently advanced a technique for time-series data, specifically in neurobiological sciences, related to some of the fundamental concepts of this work.

As trivial as it may seem, I thought it prudent to bring something to the authors' attention. Upon installation, the `annoy` dependency threw an error and exited because of it's need for Visual Studio C++ Build Tools. Now, admittedly, this wouldn't have occurred on my office or lab computers. However, I tested this package on a new laptop that I had yet to install C++ Build Tools for until my attempted installation of this package brought its absence to my attention. I suppose its need is handled in the error message of 'annoy', as I said, it may seem trivial, but I thought it best to bring even trivialities to your attention.

Congrats!
