### 2020-01-09 #1957 @VarIr
Thank you both @ryEllison and @aozorahime  for reviewing my work and your positive feedback!

@ryEllison Do you mind pointing me to a paper of yours regarding the technique you mentioned? I'm interested, because I've got some neuroscience background and, to add to the odd events, just recently started a small project about neurophysiological time series.
Also thanks for bringing the `annoy` issue to my attention. I'm mostly dealing with Linux or Mac, so Windows issues slip through more easily. Could you perhaps open an issue at https://github.com/VarIr/scikit-hubness/ and provide the error message?
