### 2020-01-13 #1957 @terrytangyuan
Thanks everyone! @VarIr At this point could you make a new release of this software that includes the changes that have resulted from this review. Then, please make an archive of the software in Zenodo/figshare/other service and update this thread with the DOI of the archive? For the Zenodo/figshare archive, please make sure that:
* The title of the archive is the same as the JOSS paper title
* That the authors of the archive are the same as the JOSS paper authors
