### 2020-01-15 #1957 @kyleniemeyer
Hi @VarIr, the article looks good to me overall, with just a few minor things that could be corrected:
-  in the first paragraph, you have what appear to be em-dashes (—) surrounded by spaces. I believe the more-correct version would be either em-dashes without surrounding spaces (`causing—for instance—overrepresentation`, or an en-dash with spaces (`causing – for instance – overrepresentation`). I prefer the former, myself.
- At the end of the summary you have a list of links that appear as numbered references, but separate from the actual list of references. I would suggest either embedding these links into the body of the text, or making them true citations to external links (in the same way you cited HubMiner and Hub-Toolbox, for example).
