### 2020-01-16 #1957 @ryEllison
@VarIr paper is currently under review. Once it's sharable, I don't mind to send it along. Also, certainly interested in hearing about what you're working on regarding neurophysiological time-series data. My email is ryan.dean.ellison@gmail.com. Perhaps we should move the conversation there...
