### 2020-01-17 #1957 @VarIr
@kyleniemeyer Thanks for pointing that out. I removed the non-essential links altogether, and embedded the links to the github repo and docs into the text body. Em-dashes now come w/o surrounding spaces.
