### 2020-01-17 c5498a1
FIX travis macos work-around
